requirements.txt
streamlit 
ffmpeg-python 
openai-whisper
transformers 
pillow 
easyocr 
torch 
torchvision 
torchaudio 
accelerate
